+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+- ReferenceProvider provides references for symbol macros.
 - DocumentHighlightProvider provides document highlighting for messages, tasks and actions.
 
 ### Improved